requirements.txt
streamlitstreamlit==1.26.0